feat: Add service type dropdown to job creation and implement job details page
